+++ extensions/mediasession/src/main/java/com/google/android/exoplayer2/ext/mediasession/TimelineQueueEditor.java
@@ -179,7 +179,7 @@ public final class TimelineQueueEditor
   @Override
   public boolean onCommand(
       Player player,
-      ControlDispatcher controlDispatcher,
+      @Deprecated ControlDispatcher controlDispatcher,
       String command,
       @Nullable Bundle extras,
       @Nullable ResultReceiver cb) {